38626: Viewing results per participant throws error.

+++ components/ILIAS/Survey/Evaluation/class.ilSurveyResultsUserTableGUI.php
@@ -81,7 +81,7 @@ class ilSurveyResultsUserTableGUI extends ilTable2GUI
         } elseif ($timeinseconds === 0) {
             return $this->lng->txt('not_available');
         } else {
-            return sprintf("%02d:%02d:%02d", ($timeinseconds / 3600), ($timeinseconds / 60) % 60, $timeinseconds % 60);
+            return sprintf("%02d:%02d:%02d", (int) ($timeinseconds / 3600), (int) ($timeinseconds / 60) % 60, (int) $timeinseconds % 60);
         }
     }
